mesh: Organizational cleanin of aggregate + streamlined field restriction helper for field dec.
